2790-rename-ErrorModal-to-InfoModal-and-change-modal-text
