--- packages/ketcher-react/src/components/ErrorModal/constants.ts
@@ -1,13 +1,0 @@
-export const pasteErrorText =
-  'This action is unavailable via menu. Please use shortcuts instead:'
-
-export const shortcuts = [
-  { hotKey: 'CTRL/Cmd + C', label: 'for paste' },
-  { hotKey: 'CTRL/Cmd + X', label: 'for cut' },
-  { hotKey: 'CTRL/Cmd + V', label: 'for paste' }
-]
-
-export const error = {
-  message: 'Error Message',
-  close: 'Close'
-}
