+++ packages/ketcher-react/src/components/InfoModal/PasteErrorModalBody/PasteErrorModalBody.module.less
