+++ packages/ketcher-react/src/components/InfoModal/PasteErrorModalBody/PasteErrorModalBody.tsx
@@ -1,0 +1,17 @@
+import { pasteErrorText, shortcut } from '../constants'
+
+import styles from './PasteErrorModalBody.module.less'
+
+export const PasteErrorModalBody = (): JSX.Element => {
+  return (
+    <div className={styles.pasteErrorModalBody}>
+      <div>{pasteErrorText}</div>
+      <div className={styles.shortcutsBlock}>
+        <div className={styles.shortcuts}>
+          <div className={styles.shortcut}>{shortcut.hotKey}</div>
+          <div>{shortcut.label}</div>
+        </div>
+      </div>
+    </div>
+  )
+}
