+++ packages/ketcher-react/src/components/InfoModal/PasteErrorModalBody/index.ts
