+++ packages/ketcher-react/src/components/index.ts
@@ -19,4 +19,4 @@ export * from './Buttons'
 export * from './StructRender'
 export * from './Input'
 export * from './Accordion'
-export * from './ErrorModal'
+export * from './InfoModal'
